requirements.txt
anyio==4.0.0
asgiref==3.5.2
attrs==22.1.0
black==22.12.0
click==8.1.3
colorama==0.4.6
dj-database-url==1.3.0
Django==4.1.4
django-cors-headers==3.13.0
django-dotenv==1.4.2
django-heroku==0.3.1
django-unfold==0.16.0
djangorestframework==3.14.0
djangorestframework-camel-case==1.3.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.25.1
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
isort==5.10.1
Jinja2==3.1.2
jsonschema==4.17.3
MarkupSafe==2.1.3
mypy-extensions==0.4.3
mysqlclient==2.2.0
pathspec==0.10.3
platformdirs==2.6.0
psycopg2==2.9.5
PyJWT==2.6.0
pyrsistent==0.19.2
python-multipart==0.0.6
pytz==2022.6
PyYAML==6.0
sniffio==1.3.0
sqladmin==0.16.0
SQLAlchemy==2.0.23
sqlalchemy-django-admin==0.0.6
sqlparse==0.4.3
starlette==0.32.0.post1
tomli==2.0.1
tzdata==2022.7
uritemplate==4.1.1
whitenoise==6.4.0
WTForms==3.1.1
typing-extensions>=4.6.0
